+++ src/main/java/com/grepp/spring/app/model/alarm/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.auth.payload;
+package com.grepp.spring.app.model.alarm.dto.request;
 
 import lombok.Data;
